requirements.txt
asgiref==3.4.1
Django==3.2.7
django-cors-headers==3.6.0
django-crispy-forms==1.12.0
django-filter==2.4.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
pkg-resources==0.0.0
pyotp==2.6.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
